Restructured layout to allow for a simpler app structure, but does introduce more sub-grids
