+++ assets/src/edit-story/components/canvas/index.js
@@ -1,4 +1,1 @@
 export { default } from './canvas';
-export { default as Meta } from './meta';
-export { default as AddPage } from './addpage';
-export { default as Carrousel } from './carrousel';
