+++ assets/src/edit-story/components/canvas/useCanvas.js
@@ -1,0 +1,15 @@
+/**
+ * WordPress dependencies
+ */
+import { useContext } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import Context from './context';
+
+function useCanvas() {
+	return useContext( Context );
+}
+
+export default useCanvas;
