+++ assets/src/edit-story/components/library/index.js
@@ -1,3 +1,1 @@
-export { default as Library } from './library';
-export { default as LibraryTabs } from './libraryTabs';
-export { default as LibraryProvider } from './libraryProvider';
+export { default } from './library';
