+++ assets/src/edit-story/constants.js
@@ -1,0 +1,7 @@
+export const HEADER_HEIGHT = 56;
+export const LIBRARY_WIDTH = 355;
+export const INSPECTOR_WIDTH = 309;
+export const CENTRAL_RIGHT_PADDING = LIBRARY_WIDTH - INSPECTOR_WIDTH;
+
+export const PAGE_WIDTH = 434;
+export const PAGE_HEIGHT = 775;
